v2.8_to_3.1/python_etl/requirements.txt
Django==1.10.5
Pillow==4.1.1
django-extensions==1.7.9
django-htmlmin==0.10.0
django-compressor==2.1.1
django-celery-results
dj-database-url==0.4.2
amqp==2.2.2
Babel==2.5.1
backports-abc==0.5
billiard==3.5.0.3
blaze==0.10.1
celery==4.1.0
flower==0.9.2
celery-tasktree==0.3.4
certifi==2017.7.27.1
cffi==1.11.2
click==6.7
contextlib2==0.5.5
datashape==0.5.2
decorator==4.1.2
Flask==0.12.2
Flask-Cors==3.0.3
futures==3.1.1
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
multipledispatch==0.4.9
networkx==1.11
numpy==1.13.3
odo==0.5.0
pandas==0.20.3
psutil==5.4.0
psycopg2==2.7.3.1
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
rabbitmq==0.2.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
toolz==0.8.2
tornado==4.5.2
vine==1.1.4
Werkzeug==0.12.2
